--- src/main/java/com/example/ecommerceweb/dto/response/ProductResponse.java
@@ -1,4 +1,0 @@
-package com.example.ecommerceweb.dto.response;
-
-public class ProductResponse {
-}
