+++ src/main/java/com/example/ecommerceweb/dto/response/order/StripeResponse.java
@@ -1,4 +1,4 @@
-package com.example.ecommerceweb.dto.response;
+package com.example.ecommerceweb.dto.response.order;
 
 import lombok.*;
