+++ src/main/java/com/example/ecommerceweb/dto/response/product/ProductResponse.java
@@ -1,0 +1,4 @@
+package com.example.ecommerceweb.dto.response.product;
+
+public class ProductResponse {
+}
