+++ src/main/java/com/example/ecommerceweb/dto/response/user/AddressResponse.java
@@ -1,4 +1,4 @@
-package com.example.ecommerceweb.dto.response;
+package com.example.ecommerceweb.dto.response.user;
 
 import lombok.*;
